api/api/requirements.txt
PyMySQL==1.0.2
python-multipart==0.0.5
fastapi==0.82.0
uvicorn==0.18.3
SQLAlchemy==1.4.32
python-jose==3.3.0
passlib==1.7.4
PyJWT==2.4.0
pandas==1.4.2
scikit-learn==1.0.2
imbalanced-learn==0.7.0
dill==0.3.5.1